Add fwup_env key for passing variables to fwup
This lets users configure OS environment variables so that they can
affect how firmware updates are applied. To use this, you escape
environment variables in the `fwup.conf` files so that they are
evaluated when the firmware update is applied rather than when it is
created.

Some things that can be done with this:

1. Pass device-specific information to affect what gets written to
   U-Boot environment variables
2. Pass secrets for file system encryption
3. Use device-specific file paths for new files going to FAT-formatted
   partitions (like the boot partition)

This is an opt-in feature for fwup. All of the official Nerves fwup.conf
files don't read anything from the OS environment so setting environment
variables here won't do anything without customization.