Support loading options from the application env
This makes it possible to have global defaults in the application
environment and then let them be overridden via the subsystem
configuration. This is more of a convenience than a necessity since it
is possible to override the subsystem specs with `:nerves_ssh`. However,
when you do that, you have to replicate what `:nerves_ssh` does with
configuring the devpath.

This adds docs to the README as well.