Fix references to mfa that should be mfargs
mfa = module, function, arity
mfargs = module, function, arguments

Dialyzer wasn't catching the use of `[any()]` in the tuple that was
typed to be `arity` (0..255) so this went unnoticed. This change just
clarifies that the user should pass arguments. I assume wasn't much of
an issue, since I think the A in MFA is frequently switched between
arguments and arity in libraries even though the `mfa()` type is pretty
clear that it's arity. MFArgs or `mfargs()` is how I've always seen the
argument version referred to when people make a distinction.